+++ src/constant/index.tsx
@@ -1,0 +1,7 @@
+import { ISelectOption } from "@soaltee-loyalty/components/atoms/Select";
+
+export const nationality: ISelectOption[] = [
+  { value: "Nepalese", label: "Nepalese" },
+  { value: "Indian", label: "Indian" },
+  { value: "Ameriacan", label: "Indian" },
+];